src/network/network.mqtt/requirements.txt
icmplib==3.0.4
paho-mqtt==2.1.0
pytest==8.1.1